+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * `k8s.io/apiextensions-apiserver` from `v0.35.0` to `v0.35.2`
   * `k8s.io/apimachinery` from `v0.35.0` to `v0.35.2`
   * `k8s.io/client-go` from `v0.35.0` to `v0.35.2`
+* [ENHANCEMENT] Add `-kubernetes.client-qps` and `-kubernetes.client-burst` flags to configure Kubernetes API client rate limiting. #385
 
 ## v0.35.0